// src/types/types.ts
export interface ProfilePreviewType {
  id: string;
  username: string;
  first_name: string;
  last_name: string;
  thumbnail_url: string;
  name?: string;
  // category: string;
}
// ...
export interface MomentType {
  moment_id: string;
  caption: string;
  date_created: string;
  moment_category: string;
  moment_url: string;
  thumbnail_url: string;
  view_count: number;
  share_count: number;
}

export interface MomentPostType extends MomentType {
  comments_count: number;
  user: ProfilePreviewType;
  selColor: string;
}
// ...
export type ShareToType =
  | 'Search'
  | 'Copy Link'
  | 'Snapchat'
  | 'Twitter'
  | 'Facebook'
  | 'Messenger'
  | 'Instagram'
  | 'Stories'
  | 'SMS'
  | 'Others';